refactor(mobile): move todos fetching into todos.service and use in hook
